+++ modules/polyfills/src/file/blob-polyfill.d.ts
@@ -1,0 +1,1 @@
+export const BlobPolyfill: typeof Blob;