+++ modules/polyfills/src/file/file-polyfill.d.ts
@@ -1,0 +1,3 @@
+export const FilePolyfill: typeof File;
+
+export const FileReaderPolyfill: typeof FileReader;